requirements.txt
asgiref==3.3.1
Django==3.1.7
django-environ==0.4.5
django-multiselectfield==0.1.12
django-taggit==1.3.0
djangorestframework==3.12.4
flake8==3.9.0
install==1.3.4
mccabe==0.6.1
Pillow==8.2.0
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.0
pytz==2021.1
reportlab==3.5.67
sorl-thumbnail==12.7.0
sqlparse==0.4.1
